Some updates

+++ app.js
@@ -55,6 +55,10 @@ log('server start');
 var io = require('socket.io').listen(require('./package.json').port, {log:false});
 
 function _broadcast(id, msg, data){
+	if (!data) {
+		data = {};
+	}
+	data.roomid = id;
 	if(DEBUG){
 		if(data){
 			log('>> [' + id + ']', msg, data);
@@ -110,7 +114,7 @@ io.sockets.on('connection', function(socket){
 	log('[' + ip + ']', 'connect "socket begin"');
 
 	function check(data){
-		if(data === undefined){
+		if((data === undefined) || (data === null)){
 			return false;
 		}
 		for(var i = 1; i < arguments.length; i++){
@@ -123,7 +127,7 @@ io.sockets.on('connection', function(socket){
 
 	socket.on('disconnect', function(){
 		if(socket.session){
-			_leave();
+			_leaveAll();
 		}
 	});
 
@@ -198,7 +202,7 @@ io.sockets.on('connection', function(socket){
 
 	socket.on('logout', function(){
 		if(socket.session){
-			_leave();
+			_leaveAll();
 			delete session[socket.session.sid];
 			delete users[socket.session.user.name];
 			delete socket.session;
@@ -283,7 +287,7 @@ io.sockets.on('connection', function(socket){
 			socket.emit('delete', {err:err});
 			if(!err && rooms[data.path]){
 				var room = rooms[data.path];
-				socket.broadcast.to(room.id).emit('deleted');
+				socket.broadcast.to(room.id).emit('deleted', {roomid:room.id});
 				for(var u in room.users){
 					users[u].leave(room.id);
 					delete users[u].session.room;
@@ -311,7 +315,7 @@ io.sockets.on('connection', function(socket){
 				delete rooms[data.path];
 				room.path = data.newPath;
 				rooms[room.path] = room;
-				socket.broadcast.to(room.id).emit('moved', {newPath:data.newPath, time:new Date().getTime()});
+				socket.broadcast.to(room.id).emit('moved', {roomid:room.id, newPath:data.newPath, time:new Date().getTime()});
 			}
 		});
 	});
@@ -330,7 +334,7 @@ io.sockets.on('connection', function(socket){
 				userDAO.getUserByName(data.name, function(err, user){
 					if(!err){
 						var room = rooms[data.path];
-						socket.broadcast.to(room.id).emit('shared', {name:user.name, avatar:user.avatar, time:new Date().getTime()});
+						socket.broadcast.to(room.id).emit('shared', {roomid: room.id, name:user.name, avatar:user.avatar, time:new Date().getTime()});
 					}
 				});
 			}
@@ -358,7 +362,7 @@ io.sockets.on('connection', function(socket){
 					}
 				}
 				if(room){
-					socket.broadcast.to(room.id).emit('unshared', {name:data.name, time:new Date().getTime()});
+					socket.broadcast.to(room.id).emit('unshared', {roomid: room.id, name:data.name, time:new Date().getTime()});
 					if(room.users[data.name]){
 						users[data.name].leave(room.id);
 						delete users[data.name].session.room;
@@ -393,12 +397,25 @@ io.sockets.on('connection', function(socket){
 		});
 	});
 
-	function _leave(){
+	function _leaveAll() {
+		if (socket.session && socket.session.rooms) {
+			var i;
+			for (i in socket.session.rooms) {
+				_leave(socket.session.rooms[i]);
+			}
+		}
+	}
+
+	function _leave(roomid){
 		if(socket.session && socket.session.room){
 			var user = socket.session.user;
-			var room = socket.session.room;
+			var room = socket.session.rooms[roomid];
+			if (!room) {
+				return;
+			}
 			socket.leave(room.id);
-			delete socket.session.room;
+			//delete socket.session.room;
+			delete socket.session.rooms[roomid];
 			_broadcast(room.id, 'leave', {name:user.name, time:new Date().getTime()});
 			delete room.users[user.name];
 			room.count--;
@@ -423,7 +440,7 @@ io.sockets.on('connection', function(socket){
 			return socket.emit('unauthorized');
 		}
 		var user = socket.session.user;
-		_leave();
+		//_leave();
 		docDAO.getRevision(user._id, data.path, 0, function(err, revision){
 			if(err){
 				return socket.emit('join', {err:err});
@@ -432,11 +449,15 @@ io.sockets.on('connection', function(socket){
 			if(!room){
 				room = rooms[data.path] = {id:revision.doc, path:data.path, count:0, users:{}, version:0, buffer:new DocBuffer(revision.content), bps:'', exprs:{}};	
 			}else{
-				socket.broadcast.to(room.id).emit('join', {name:user.name, time:new Date().getTime()});
+				socket.broadcast.to(room.id).emit('join', {roomid:room.id, name:user.name, time:new Date().getTime()});
 			}
 			room.users[user.name] = true;
 			room.count++;
-			socket.session.room = room;
+			if (!socket.session.rooms) {
+				socket.session.rooms = {};
+			}
+			socket.session.rooms[room.id] = room;
+			//socket.session.room = room;
 			socket.join(room.id);
 			var r = {id:room.id, users:room.users, version:room.version, text:room.buffer.toString(), bps:room.bps, exprs:room.exprs};
 			if(room.runner){
@@ -456,23 +477,24 @@ io.sockets.on('connection', function(socket){
 		if(!socket.session){
 			return socket.emit('unauthorized');
 		}
-		_leave();
+		_leave(data.path);
+		//_leave();
 	});
 
 	socket.on('change', function(data){ // version, from, to, text
-		if(!check(data, 'version', 'from', 'to', 'text')){
+		if(!check(data, 'roomid', 'version', 'from', 'to', 'text')){
 			return;
 		}
 		if(!socket.session){
 			return socket.emit('unauthorized');
 		}
 		var user = socket.session.user;
-		var room = socket.session.room;
+		var room = socket.session.rooms[data.roomid];
 		if(room && room.version == data.version && !room.dbger){
 			room.version = (room.version + 1) % 65536;
 			room.buffer.update(data.from, data.to, data.text, function(err){
 				if(!err){
-					socket.emit('ok');
+					socket.emit('ok', {roomid: data.roomid});
 					data.name = user.name;
 					socket.broadcast.to(room.id).emit('change', data);
 				}
@@ -481,21 +503,21 @@ io.sockets.on('connection', function(socket){
 	});
 
 	socket.on('bps', function(data){ // version, from, to, text
-		if(!check(data, 'version', 'from', 'to', 'text')){
+		if(!check(data, 'roomid', 'version', 'from', 'to', 'text')){
 			return;
 		}
 		if(!socket.session){
 			return socket.emit('unauthorized');
 		}
 		var user = socket.session.user;
-		var room = socket.session.room;
+		var room = socket.session.rooms[data.roomid];
 		if(room && room.version == data.version && data.from <= data.to && data.from >= 0){
 			room.version = (room.version + 1) % 65536;
 			while(room.bps.length < data.to){
 				room.bps += '0';
 			}
 			room.bps = room.bps.substr(0, data.from) + data.text + room.bps.substr(data.to);
-			socket.emit('bpsok');
+			socket.emit('bpsok', {roomid: data.roomid});
 			data.name = user.name;
 			socket.broadcast.to(room.id).emit('bps', data);
 			if(room.dbger && room.dbger.state == 'waiting'){
@@ -532,14 +554,14 @@ io.sockets.on('connection', function(socket){
 	});
 
 	socket.on('chat', function(data){ // text
-		if(!check(data, 'text')){
+		if(!check(data, 'roomid', 'text')){
 			return;
 		}
 		if(!socket.session){
 			return socket.emit('unauthorized');
 		}
 		var user = socket.session.user;
-		var room = socket.session.room;
+		var room = socket.session.rooms[data.roomid];
 		if(room){
 			data.name = user.name;
 			data.time = new Date().getTime();
@@ -548,14 +570,14 @@ io.sockets.on('connection', function(socket){
 	});
 
 	socket.on('run', function(data){ // type, version
-		if(!check(data, 'type', 'version')){
+		if(!check(data, 'roomid', 'type', 'version')){
 			return;
 		}
 		if(!socket.session){
 			return socket.emit('unauthorized');
 		}
 		var user = socket.session.user;
-		var room = socket.session.room;
+		var room = socket.session.rooms[data.roomid];
 		if(room && !room.runner && !room.dbger && room.version == data.version){
 			var runner = new Runner(room.path.substr(room.path.lastIndexOf('/') + 1), data.type, room.buffer.toString());
 			if(runner.ready()){
@@ -579,12 +601,15 @@ io.sockets.on('connection', function(socket){
 		}
 	});
 
-	socket.on('kill', function(){
+	socket.on('kill', function(data){
+		if (!check(data, 'roomid')) {
+			return;
+		}
 		if(!socket.session){
 			return socket.emit('unauthorized');
 		}
 		var user = socket.session.user;
-		var room = socket.session.room;
+		var room = socket.session.rooms[data.roomid];
 		if(room){
 			if(room.runner){
 				room.runner.kill();
@@ -595,14 +620,14 @@ io.sockets.on('connection', function(socket){
 	});
 
 	socket.on('stdin', function(data){ // data
-		if(!check(data, 'data')){
+		if(!check(data, 'roomid', 'data')){
 			return;
 		}
 		if(!socket.session){
 			return socket.emit('unauthorized');
 		}
 		var user = socket.session.user;
-		var room = socket.session.room;
+		var room = socket.session.rooms[data.roomid];
 		if(room){
 			var t;
 			if(room.runner){
@@ -613,7 +638,7 @@ io.sockets.on('connection', function(socket){
 			if(t){
 				t.input(data.data, function(err){
 					if(err){
-						return socket.emit('stdin', {err:err});
+						return socket.emit('stdin', {roomid:room.id, err:err});
 					}
 					_broadcast(room.id, 'stdin', data);
 				});
@@ -622,14 +647,14 @@ io.sockets.on('connection', function(socket){
 	});
 
 	socket.on('debug', function(data){ // type, version
-		if(!check(data, 'type', 'version')){
+		if(!check(data, 'roomid', 'type', 'version')){
 			return;
 		}
 		if(!socket.session){
 			return socket.emit('unauthorized');
 		}
 		var user = socket.session.user;
-		var room = socket.session.room;
+		var room = socket.session.rooms[data.roomid];
 		if(room && !room.runner && !room.dbger && room.version == data.version){
 			var dbger = new Debugger(room.path.substr(room.path.lastIndexOf('/') + 1), data.type, room.buffer.toString());
 			if(dbger.ready()){
@@ -694,59 +719,71 @@ io.sockets.on('connection', function(socket){
 		}
 	});
 
-	socket.on('step', function(){
-		if(!socket.session){
-			return socket.emit('unauthorized');
-		}
-		var user = socket.session.user;
-		var room = socket.session.room;
-		if(room && room.dbger){
-			return room.dbger.step();
-		}
-	});
-
-	socket.on('next', function(){
-		if(!socket.session){
-			return socket.emit('unauthorized');
-		}
-		var user = socket.session.user;
-		var room = socket.session.room;
-		if(room && room.dbger){
-			return room.dbger.next();
-		}
-	});
-
-	socket.on('resume', function(){
-		if(!socket.session){
-			return socket.emit('unauthorized');
-		}
-		var user = socket.session.user;
-		var room = socket.session.room;
-		if(room && room.dbger){
-			return room.dbger.resume();
-		}
-	});
-
-	socket.on('finish', function(){
-		if(!socket.session){
-			return socket.emit('unauthorized');
-		}
-		var user = socket.session.user;
-		var room = socket.session.room;
-		if(room && room.dbger){
-			return room.dbger.finish();
-		}
-	});
-
-	socket.on('add-expr', function(data){ // expr
-		if(!check(data, 'expr')){
+	socket.on('step', function(data){
+		if (!check(data, 'roomid')) {
 			return;
 		}
 		if(!socket.session){
 			return socket.emit('unauthorized');
 		}
 		var user = socket.session.user;
-		var room = socket.session.room;
+		var room = socket.session.rooms[data.roomid];
+		if(room && room.dbger){
+			return room.dbger.step();
+		}
+	});
+
+	socket.on('next', function(data){
+		if (!check(data, 'roomid')) {
+			return;
+		}
+		if(!socket.session){
+			return socket.emit('unauthorized');
+		}
+		var user = socket.session.user;
+		var room = socket.session.rooms[data.roomid];
+		if(room && room.dbger){
+			return room.dbger.next();
+		}
+	});
+
+	socket.on('resume', function(data){
+		if (!check(data, 'roomid')) {
+			return;
+		}
+		if(!socket.session){
+			return socket.emit('unauthorized');
+		}
+		var user = socket.session.user;
+		var room = socket.session.rooms[data.roomid];
+		if(room && room.dbger){
+			return room.dbger.resume();
+		}
+	});
+
+	socket.on('finish', function(data){
+		if (!check(data, 'roomid')) {
+			return;
+		}
+		if(!socket.session){
+			return socket.emit('unauthorized');
+		}
+		var user = socket.session.user;
+		var room = socket.session.rooms[data.roomid];
+		if(room && room.dbger){
+			return room.dbger.finish();
+		}
+	});
+
+	socket.on('add-expr', function(data){ // expr
+		if(!check(data, 'roomid', 'expr')){
+			return;
+		}
+		if(!socket.session){
+			return socket.emit('unauthorized');
+		}
+		var user = socket.session.user;
+		var room = socket.session.rooms[data.roomid];
 		if(room && !room.exprs.hasOwnProperty(data.expr) && data.expr != ''){
 			room.exprs[data.expr] = null;
 			if(room.dbger && room.dbger.state == 'waiting'){
@@ -764,14 +801,14 @@ io.sockets.on('connection', function(socket){
 	});
 
 	socket.on('rm-expr', function(data){ // expr
-		if(!check(data, 'expr')){
+		if(!check(data, 'roomid', 'expr')){
 			return;
 		}
 		if(!socket.session){
 			return socket.emit('unauthorized');
 		}
 		var user = socket.session.user;
-		var room = socket.session.room;
+		var room = socket.session.rooms[data.roomid];
 		if(room && room.exprs.hasOwnProperty(data.expr)){
 			delete room.exprs[data.expr];
 			return _broadcast(room.id, 'rm-expr', {expr:data.expr});
